requirements.txt
ultralytics>=8.2.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
Pillow>=10.0.0
pytest>=7.0.0
